+++ libs/cart-customer/src/public_api.ts
@@ -1,1 +1,1 @@
-export * from './injection-tokens/public_api';
+export {};
