+++ libs/cart-customer/state/src/public_api.ts
@@ -1,1 +1,2 @@
 export { DaffCartCustomerStateModule } from './state.module';
+export * from './injection-tokens/public_api';
